Edit: moved shared entities to domain layer
