+++ src/modules/authentication/repositories/IUsersRepository.ts
@@ -1,7 +1,7 @@
-import IUser from 'shared/entities/IUser';
-import IUserSignupDTO from 'modules/authentication/dtos/IUserSignupDTO';
 import IUserLoginDTO from 'modules/authentication/dtos/IUserLoginDTO';
+import IUserSignupDTO from 'modules/authentication/dtos/IUserSignupDTO';
 import IUserAuthentication from 'modules/authentication/entities/IUserAuthentication';
+import IUser from 'shared/domain/entities/IUser';
 
 export default interface IUsersRepository {
   create(values: IUserSignupDTO): Promise<IUser>;
