+++ src/modules/chooseGame/infra/controllers/useGetUserController.ts
@@ -2,7 +2,7 @@ import makeGetUserService from 'modules/chooseGame/services/factories/makeGetUse
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
 import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
-import IUser from 'shared/entities/IUser';
+import IUser from 'shared/domain/entities/IUser';
 
 interface UserParams {
   userId: string;
