+++ src/modules/chooseGame/repositories/IGamesRepository.ts
@@ -1,4 +1,4 @@
-import IGame from 'shared/entities/IGame';
+import IGame from 'shared/domain/entities/IGame';
 
 export default interface IGamesRepository {
   findById(id: string): Promise<IGame>;
