+++ src/modules/chooseGame/repositories/IPlayersRepository.ts
@@ -1,4 +1,4 @@
-import IPlayer from 'shared/entities/IPlayer';
+import IPlayer from 'shared/domain/entities/IPlayer';
 
 export default interface IPlayersRepository {
   create(gameId: string): Promise<void>;
