+++ src/modules/chooseGame/services/FetchPlayersService.ts
@@ -1,5 +1,5 @@
 import IPlayersRepository from 'modules/chooseGame/repositories/IPlayersRepository';
-import IPlayer from 'shared/entities/IPlayer';
+import IPlayer from 'shared/domain/entities/IPlayer';
 
 interface IExecute {
   players?: IPlayer[];
