+++ src/modules/chooseGame/services/GetGameService.ts
@@ -1,5 +1,5 @@
 import IGamesRepository from 'modules/chooseGame/repositories/IGamesRepository';
-import IGame from 'shared/entities/IGame';
+import IGame from 'shared/domain/entities/IGame';
 
 interface IExecute {
   game?: IGame;
