+++ src/modules/selectedGame/container/contexts/PlayerProfileContext/models/IPlayerProfileContext.ts
@@ -1,4 +1,4 @@
-import IPlayer from 'shared/entities/IPlayer';
+import IPlayer from 'shared/domain/entities/IPlayer';
 
 export default interface IPlayerProfileContext {
   player: IPlayer;
