+++ src/modules/selectedGame/repositories/IPlayersRepository.ts
@@ -1,4 +1,4 @@
-import IPlayer from 'shared/entities/IPlayer';
+import IPlayer from 'shared/domain/entities/IPlayer';
 
 export default interface IPlayersRepository {
   findById(id: string): Promise<IPlayer>;
