+++ src/modules/selectedGame/view/pages/Feed/components/RankTag/index.tsx
@@ -1,12 +1,7 @@
 import React from 'react';
 
-// Components
 import { RankChip } from 'modules/selectedGame/view/components';
-
-// Entities
-import IRank from 'shared/entities/IRank';
-
-// Helpers
+import IRank from 'shared/domain/entities/IRank';
 import { getTextColor } from 'shared/view/helpers';
 
 interface RankTagProps {
