+++ src/shared/domain/entities/ILevelInfo.ts
