+++ src/shared/domain/entities/IRank.ts
