+++ src/shared/domain/entities/IUser.ts
